generator/requirements.txt
tensorflow
keras
numpy
kafka-python
